move metadata file
