+++ src/io/parquet/mod.rs
@@ -33,10 +33,11 @@
 //! }
 //! ```
 
-mod geoparquet_metadata;
+mod metadata;
 mod reader;
 #[cfg(feature = "parquet_async")]
 mod reader_async;
+mod writer;
 
 pub use reader::{read_geoparquet, GeoParquetReaderOptions};
 #[cfg(feature = "parquet_async")]
