+++ src/io/parquet/reader_async.rs
@@ -1,5 +1,5 @@
 use crate::error::Result;
-use crate::io::parquet::geoparquet_metadata::build_arrow_schema;
+use crate::io::parquet::metadata::build_arrow_schema;
 use crate::io::parquet::reader::GeoParquetReaderOptions;
 use crate::table::GeoTable;
